Add mixins folder
This moves several files to a new `mixins/` folder. I was going to add
some new macros, per [this comment][1], but thought it might clutter
things up a bit to put them in the root folder. I thought of adding a
`macros/` folder, but decided to emulate the [rubocop source][2]
instead.

[1]: https://github.com/rubocop-hq/rubocop-rspec/pull/934#issuecomment-643813427
[2]: https://github.com/rubocop-hq/rubocop/tree/master/lib/rubocop/cop/mixin
