+++ lib/rubocop/cop/rspec/instance_variable.rb
@@ -47,7 +47,7 @@ module RuboCop
       #   end
       #
       class InstanceVariable < Base
-        include RuboCop::RSpec::TopLevelGroup
+        include TopLevelGroup
 
         MSG = 'Avoid instance variables – use let, ' \
               'a method call, or a local variable (if possible).'
