+++ lib/rubocop/cop/rspec/subject_stub.rb
@@ -22,7 +22,7 @@ module RuboCop
       #   end
       #
       class SubjectStub < Base
-        include RuboCop::RSpec::TopLevelGroup
+        include TopLevelGroup
 
         MSG = 'Do not stub methods of the object under test.'
